feat: distributions update within cart after adding a course from cart
