+++ lib/components/dashboard/degree-info/CourseBar.tsx
@@ -4,6 +4,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectCurrentPlanCourses,
   selectDistributions,
+  updateSelectedDistribution,
 } from '../../../slices/currentPlanSlice';
 import { requirements } from './distributionFunctions';
 import { CheckCircleIcon, ExclamationIcon } from '@heroicons/react/solid';
@@ -11,7 +12,6 @@ import ReactTooltip from 'react-tooltip';
 import {
   updateAddingPrereq,
   updateInfoPopup,
-  updateSelectedDistribution,
   updateShowingCart,
 } from '../../../slices/popupSlice';
 import { clearSearch, updatePlaceholder } from '../../../slices/searchSlice';
